cleanup(angular): remove deprecated syntax option from ngrx generator
