+++ packages/angular/src/generators/ngrx/files/__directory__/__fileName__.actions.ts__tmpl__
