+++ packages/angular/src/generators/ngrx/files/__directory__/__fileName__.effects.spec.ts__tmpl__
