+++ packages/angular/src/generators/ngrx/files/__directory__/__fileName__.facade.spec.ts__tmpl__
