+++ packages/angular/src/generators/ngrx/files/__directory__/__fileName__.facade.ts__tmpl__
