+++ packages/angular/src/generators/ngrx/files/__directory__/__fileName__.models.ts__tmpl__
