+++ packages/angular/src/generators/ngrx/files/__directory__/__fileName__.reducer.spec.ts__tmpl__
