+++ packages/angular/src/generators/ngrx/files/__directory__/__fileName__.reducer.ts__tmpl__
