+++ packages/angular/src/generators/ngrx/files/__directory__/__fileName__.selectors.spec.ts__tmpl__
