+++ packages/angular/src/generators/ngrx/files/__directory__/__fileName__.selectors.ts__tmpl__
