--- packages/angular/src/generators/ngrx/files/classes-syntax/__directory__/__fileName__.reducer.ts__tmpl__
@@ -1,47 +1,0 @@
-import { <%= className %>Action, <%= className %>ActionTypes } from './<%= fileName %>.actions';
-
-export const <%= constantName %>_FEATURE_KEY = '<%= propertyName %>';
-
-/**
- * Interface for the '<%= className %>' data used in
- *  - <%= className %>State, and the reducer function
- *
- *  Note: replace if already defined in another module
- */
-export interface <%= className %>Entity {
-  id: string;
-  name: string;
-}
-
-export interface <%= className %>State {
-  list: <%= className %>Entity[]; // list of <%= className %>; analogous to a sql normalized table
-  selectedId?: string | number; // which <%= className %> record has been selected
-  loaded: boolean; // has the <%= className %> list been loaded
-  error?: any; // last none error (if any)
-};
-
-export interface <%= className %>PartialState {
-  readonly [<%= constantName %>_FEATURE_KEY]: <%= className %>State;
-}
-
-export const initial<%= className %>State: <%= className %>State = {
-  list: [],
-  loaded: false
-};
-
-export function <%= propertyName %>Reducer(
-  state: <%= className %>State = initial<%= className %>State,
-  action: <%= className %>Action): <%= className %>State
-{
-  switch (action.type) {
-    case <%= className %>ActionTypes.<%= className %>Loaded: {
-      state = {
-        ...state,
-        list: action.payload,
-        loaded: true
-      };
-      break;
-    }
-  }
-  return state;
-}
