+++ packages/angular/src/generators/ngrx/schema.json
@@ -57,13 +57,6 @@
       "default": false,
       "description": "Do not update the `package.json` with NgRx dependencies."
     },
-    "syntax": {
-      "type": "string",
-      "enum": ["classes", "creators"],
-      "default": "creators",
-      "description": "Specifies whether to use class-based or creator functions for actions, reducers, and effects.",
-      "x-deprecated": "Classes syntax is deprecated and this option will be removed in v15. Creators syntax will be the default in v15 and this option will be removed."
-    },
     "useDataPersistence": {
       "type": "boolean",
       "default": false,
